refactor(testing): consolidate createTestContainer into core
- Remove resolveWith from public API
- Move createTestContainer to @zeltjs/core
- Testing package re-exports from core (deprecation path)
