+++ packages/core/src/index.ts
@@ -55,8 +55,8 @@ export type {
 export { Config, injectConfig } from './config';
 export type { ConfigClass } from './config';
 
-export { resolveWith } from './internal/container';
-export type { ResolverHandle } from './internal/container';
+export { createTestContainer } from './internal/container';
+export type { CreateTestContainerOptions, TestContainerResult } from './internal/container';
 
 export { LifecycleManager } from './lifecycle';
 export type { Disposable } from './lifecycle';
