+++ packages/testing/src/index.ts
@@ -1,5 +1,2 @@
 export { createTestContainer } from './test-container';
 export type { CreateTestContainerOptions, TestContainerResult } from './test-container';
-
-export { resolveWith } from '@zeltjs/core';
-export type { ResolverHandle } from '@zeltjs/core';
